docs(core): fix typos and improve documentation clarity

+++ src/main/java/org/apache/ddlutils/PlatformInfo.java
@@ -31,7 +31,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 /**
- * Conatains information about the database platform such as supported features and native type mappings.
+ * Contains information about the database platform such as supported features and native type mappings.
  *
  * @version $Revision: 289996 $
  */
@@ -205,7 +205,7 @@ public class PlatformInfo {
    */
   private int _maxForeignKeyNameLength = -1;
   /**
-   * The string used for delimiting SQL identifiers, eg. table names, column names etc.
+   * The string used for delimiting SQL identifiers, e.g. table names, column names etc.
    */
   private String _delimiterToken = "\"";
   /**
@@ -221,7 +221,7 @@ public class PlatformInfo {
    */
   private String _commentSuffix = "";
   /**
-   * The text separating individual sql commands.
+   * The text separating individual SQL commands.
    */
   private String _sqlCommandDelimiter = ";";
   /**
@@ -302,8 +302,8 @@ public class PlatformInfo {
   // properties influencing the specification of table constraints
 
   /**
-   * Determines whether primary key constraints are embedded in the create
-   * table clause or as seperate alter table statements. The default is
+   * Determines whether primary key constraints are embedded in the creation
+   * table clause or as separate alter table statements. The default is
    * embedded pks.
    *
    * @return <code>true</code> if pk constraints are embedded
@@ -313,8 +313,8 @@ public class PlatformInfo {
   }
 
   /**
-   * Specifies whether the primary key constraints are embedded in the create
-   * table clause or as seperate alter table statements.
+   * Specifies whether the primary key constraints are embedded in the creation
+   * table clause or as separate alter table statements.
    *
    * @param primaryKeyEmbedded Whether pk constraints are embedded
    */
@@ -343,7 +343,7 @@ public class PlatformInfo {
   /**
    * Determines whether primary key can contain identity and non-identity columns at the same time.
    *
-   * @return <code>true</code> if idenity and non-identity columns can be mixed in the pk
+   * @return <code>true</code> if identity and non-identity columns can be mixed in the pk
    */
   public boolean isMixingIdentityAndNormalPrimaryKeyColumnsSupported() {
     return _mixingIdentityAndNormalPrimaryKeyColumnsSupported;
@@ -352,7 +352,7 @@ public class PlatformInfo {
   /**
    * Specifies whether primary key can contain identity and non-identity columns at the same time.
    *
-   * @param mixingIdentityAndNormalPrimaryKeyColumnsSupported Whether idenity and non-identity columns
+   * @param mixingIdentityAndNormalPrimaryKeyColumnsSupported Whether identity and non-identity columns
    *                                                          can be mixed in the pk
    */
   public void setMixingIdentityAndNormalPrimaryKeyColumnsSupported(boolean mixingIdentityAndNormalPrimaryKeyColumnsSupported) {
@@ -371,8 +371,8 @@ public class PlatformInfo {
   }
 
   /**
-   * Specifies whether foreign key constraints are embedded in the create
-   * table clause or as seperate alter table statements.
+   * Specifies whether foreign key constraints are embedded in the creation
+   * table clause or as separate alter table statements.
    *
    * @param foreignKeysEmbedded Whether fk constraints are embedded
    */
@@ -417,8 +417,8 @@ public class PlatformInfo {
   }
 
   /**
-   * Determines whether the indices are embedded in the create table clause
-   * or as seperate statements. Per default, indices are external.
+   * Determines whether the indices are embedded in the creation table clause
+   * or as separate statements. Per default, indices are external.
    *
    * @return <code>true</code> if indices are embedded
    */
@@ -427,8 +427,8 @@ public class PlatformInfo {
   }
 
   /**
-   * Specifies whether indices are embedded in the create table clause or
-   * as seperate alter table statements.
+   * Specifies whether indices are embedded in the creation table clause or
+   * as separate alter table statements.
    *
    * @param indicesEmbedded Whether indices are embedded
    */
@@ -517,7 +517,7 @@ public class PlatformInfo {
   }
 
   /**
-   * Determines whether the database will make an idenity column automatically required.
+   * Determines whether the database will make an identity column automatically required.
    *
    * @return <code>true</code> if identity columns are automatically required
    */
@@ -681,7 +681,7 @@ public class PlatformInfo {
   }
 
   /**
-   * Determines whether the platform is allows the explicit specification of values for
+   * Determines whether the platform is allowing the explicit specification of values for
    * identity columns in INSERT/UPDATE statements.
    *
    * @return <code>true</code> if values for identity columns can be specified
@@ -691,7 +691,7 @@ public class PlatformInfo {
   }
 
   /**
-   * Specifies whether the platform is allows the explicit specification of values for
+   * Specifies whether the platform is allowing the explicit specification of values for
    * identity columns in INSERT/UPDATE statements.
    *
    * @param identityOverrideAllowed <code>true</code> if values for identity columns can be specified
@@ -829,7 +829,7 @@ public class PlatformInfo {
   }
 
   /**
-   * Returns the text that is used to delimit identifiers (eg. table names).
+   * Returns the text that is used to delimit identifiers (e.g. table names).
    * Per default, this is a double quotation character (").
    *
    * @return The delimiter text
@@ -839,7 +839,7 @@ public class PlatformInfo {
   }
 
   /**
-   * Sets the text that is used to delimit identifiers (eg. table names).
+   * Sets the text that is used to delimit identifiers (e.g. table names).
    *
    * @param delimiterToken The delimiter text
    */
@@ -848,7 +848,7 @@ public class PlatformInfo {
   }
 
   /**
-   * Returns the text that is used for for quoting values (e.g. text) when
+   * Returns the text that is used for quoting values (e.g. text) when
    * printing default values and in generates insert/update/delete statements.
    * Per default, this is a single quotation character (').
    *
@@ -859,7 +859,7 @@ public class PlatformInfo {
   }
 
   /**
-   * Sets the text that is used for for quoting values (e.g. text) when
+   * Sets the text that is used for quoting values (e.g. text) when
    * printing default values and in generates insert/update/delete statements.
    *
    * @param valueQuoteChar The new quote text
@@ -906,7 +906,7 @@ public class PlatformInfo {
   }
 
   /**
-   * Returns the text separating individual sql commands.
+   * Returns the text separating individual SQL commands.
    *
    * @return The delimiter text
    */
@@ -915,7 +915,7 @@ public class PlatformInfo {
   }
 
   /**
-   * Sets the text separating individual sql commands.
+   * Sets the text separating individual SQL commands.
    *
    * @param sqlCommandDelimiter The delimiter text
    */
@@ -946,7 +946,7 @@ public class PlatformInfo {
   public int getTargetJdbcType(int typeCode) {
     Integer targetJdbcType = _targetJdbcTypes.get(typeCode);
 
-    return targetJdbcType == null ? typeCode : targetJdbcType.intValue();
+    return targetJdbcType == null ? typeCode : targetJdbcType;
   }
 
   /**
@@ -1020,10 +1020,10 @@ public class PlatformInfo {
   }
 
   /**
-   * Determines whether the native type for the given sql type code (one of the
+   * Determines whether the native type for the given SQL type code (one of the
    * {@link java.sql.Types} constants) has a null default value on this platform.
    *
-   * @param sqlTypeCode The sql type code
+   * @param sqlTypeCode The SQL type code
    * @return <code>true</code> if the native type has a null default value
    */
   public boolean hasNullDefault(int sqlTypeCode) {
@@ -1031,10 +1031,10 @@ public class PlatformInfo {
   }
 
   /**
-   * Specifies whether the native type for the given sql type code (one of the
+   * Specifies whether the native type for the given SQL type code (one of the
    * {@link java.sql.Types} constants) has a null default value on this platform.
    *
-   * @param sqlTypeCode    The sql type code
+   * @param sqlTypeCode    The SQL type code
    * @param hasNullDefault <code>true</code> if the native type has a null default value
    */
   public void setHasNullDefault(int sqlTypeCode, boolean hasNullDefault) {
@@ -1046,10 +1046,10 @@ public class PlatformInfo {
   }
 
   /**
-   * Determines whether the native type for the given sql type code (one of the
+   * Determines whether the native type for the given SQL type code (one of the
    * {@link java.sql.Types} constants) has a size specification on this platform.
    *
-   * @param sqlTypeCode The sql type code
+   * @param sqlTypeCode The SQL type code
    * @return <code>true</code> if the native type has a size specification
    */
   public boolean hasSize(int sqlTypeCode) {
@@ -1057,10 +1057,10 @@ public class PlatformInfo {
   }
 
   /**
-   * Specifies whether the native type for the given sql type code (one of the
+   * Specifies whether the native type for the given SQL type code (one of the
    * {@link java.sql.Types} constants) has a size specification on this platform.
    *
-   * @param sqlTypeCode The sql type code
+   * @param sqlTypeCode The SQL type code
    * @param hasSize     <code>true</code> if the native type has a size specification
    */
   public void setHasSize(int sqlTypeCode, boolean hasSize) {
@@ -1109,11 +1109,11 @@ public class PlatformInfo {
   }
 
   /**
-   * Determines whether the native type for the given sql type code (one of the
+   * Determines whether the native type for the given SQL type code (one of the
    * {@link java.sql.Types} constants) has precision and scale specifications on
    * this platform.
    *
-   * @param sqlTypeCode The sql type code
+   * @param sqlTypeCode The SQL type code
    * @return <code>true</code> if the native type has precision and scale specifications
    */
   public boolean hasPrecisionAndScale(int sqlTypeCode) {
@@ -1121,11 +1121,11 @@ public class PlatformInfo {
   }
 
   /**
-   * Specifies whether the native type for the given sql type code (one of the
+   * Specifies whether the native type for the given SQL type code (one of the
    * {@link java.sql.Types} constants) has precision and scale specifications on
    * this platform.
    *
-   * @param sqlTypeCode          The sql type code
+   * @param sqlTypeCode          The SQL type code
    * @param hasPrecisionAndScale <code>true</code> if the native type has precision and scale specifications
    */
   public void setHasPrecisionAndScale(int sqlTypeCode, boolean hasPrecisionAndScale) {
@@ -1238,7 +1238,7 @@ public class PlatformInfo {
   }
 
   /**
-   * Determiones whether the two ON UPDATE actions are equivalent. Equivalent actions will not
+   * Determines whether the two ON UPDATE actions are equivalent. Equivalent actions will not
    * cause a foreign key to be changed/recreated when altering a database.
    *
    * @param actionA The first action
@@ -1277,7 +1277,7 @@ public class PlatformInfo {
   }
 
   /**
-   * Determiones whether the two ON DELETE actions are equivalent. Equivalent actions will not
+   * Determines whether the two ON DELETE actions are equivalent. Equivalent actions will not
    * cause a foreign key to be changed/recreated when altering a database.
    *
    * @param actionA The first action
